Add Grandmaster

+++ BusinessLogic/DeleteUser.cs
@@ -1,4 +1,5 @@
-﻿using Data;
+﻿using Cache;
+using Data;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -7,7 +8,7 @@ using System.Threading.Tasks;
 
 namespace BusinessLogic
 {
-   public class DeleteUser
+    public class DeleteUser
     {
         public int Id { get; set; }
         public DeleteUser(int id)
@@ -25,7 +26,13 @@ namespace BusinessLogic
 
                 model = db.Users.Where(x => x.UserId == this.Id).FirstOrDefault();
 
-                if (model.Position != "admin")
+                if (ActiveUser.c_Position == "GM")
+                {
+                    db.Users.Remove(model);
+                    db.SaveChanges();
+                }
+
+                if (model.Position != "admin" && ActiveUser.c_Position != "GM")
                 {
                     db.Users.Remove(model);
                     db.SaveChanges();
@@ -34,6 +41,9 @@ namespace BusinessLogic
                 {
                     throw new Exception("User is admin");
                 }
+
+
+
             }
 
         }
